pyproject.toml
[tool.poetry]
name = "pie"
version = "0.1.0"
description = "Extracts text from PDF and searches for IOCs"
authors = ["Mitch Haddadi <12067255+dfirsec@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pdfminer-six = "^20240706"
requests = "^2.32.3"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
